+++ services/blockchain-indexer/methods/dataService/controllers/dposVotesReceived.js
@@ -26,8 +26,8 @@ const {
 
 const getVoters = async params => {
 	const isFound = await confirmAnyId(params);
-	if (!isFound && params.address) return { status: NOT_FOUND, data: { error: `Account ${params.address} not found.` } };
-	if (!isFound && params.name) return { status: NOT_FOUND, data: { error: `Account ${params.name} not found.` } };
+	if (!isFound && params.address) return { status: NOT_FOUND, data: { error: `Account with address ${params.address} not found.` } };
+	if (!isFound && params.name) return { status: NOT_FOUND, data: { error: `Account with name ${params.name} not found.` } };
 
 	const response = await dataService.getVoters(params);
